remove extra lines in chapter 2
Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

On branch master
Changes to be committed:
	modified:   chapter2.txt

Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

On branch master
Changes to be committed:
	modified:   chapter2.txt

Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

On branch master
Changes to be committed:
	modified:   chapter2.txt

Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

On branch master
Changes to be committed:
	modified:   chapter2.txt